requirements.txt
tensorflow-cpu==2.9.0
Flask==2.2.2
gunicorn==20.1.0
Flask-Cors==3.0.10
Pillow==9.2.0